Added Bomb class, plant_bomb-method, and timer
