+++ app/const.py
@@ -1,0 +1,3 @@
+SCREEN_WIDTH = 650
+SCREEN_HEIGHT = 650
+DEFAULT_OBJECT_SIZE = 50
