+++ app/main.py
@@ -1,103 +1,28 @@
 import pygame
 
 from pygame.locals import (
-    RLEACCEL,
-    K_UP,
-    K_DOWN,
-    K_LEFT,
-    K_RIGHT,
     K_ESCAPE,
     KEYDOWN,
     QUIT,
 )
 
+from app.const import SCREEN_WIDTH, SCREEN_HEIGHT
+from app.player import Player
+from app.wall import Wall
+
 pygame.init()
 clock = pygame.time.Clock()
 
-
-SCREEN_WIDTH = 650
-SCREEN_HEIGHT = 650
-DEFAULT_OBJECT_SIZE = 50
-
 screen = pygame.display.set_mode([SCREEN_WIDTH, SCREEN_HEIGHT])
-
-
-class Player(pygame.sprite.Sprite):
-    def __init__(self):
-        super(Player, self).__init__()
-        self.surf = pygame.Surface((20, 40))
-        self.surf = pygame.image.load("images/player_front.png").convert_alpha()
-        self.rect = self.surf.get_rect()
-
-    def update(self, pressed_keys):
-        if pressed_keys[K_UP]:
-            self.rect.move_ip(0, -2)
-            if pygame.sprite.spritecollideany(player, walls):
-                self.rect.move_ip(0, 2)
-            self.surf = pygame.image.load("images/player_back.png").convert_alpha()
-        if pressed_keys[K_DOWN]:
-            self.rect.move_ip(0, 2)
-            if pygame.sprite.spritecollideany(player, walls):
-                self.rect.move_ip(0, -2)
-            self.surf = pygame.image.load("images/player_front.png").convert_alpha()
-        if pressed_keys[K_LEFT]:
-            self.rect.move_ip(-2, 0)
-            if pygame.sprite.spritecollideany(player, walls):
-                self.rect.move_ip(2, 0)
-            self.surf = pygame.image.load("images/player_left.png").convert_alpha()
-        if pressed_keys[K_RIGHT]:
-            self.rect.move_ip(2, 0)
-            if pygame.sprite.spritecollideany(player, walls):
-                self.rect.move_ip(-2, 0)
-            self.surf = pygame.image.load("images/player_right.png").convert_alpha()
-
-        if self.rect.left < 0:
-            self.rect.left = 0
-        if self.rect.right > SCREEN_WIDTH:
-            self.rect.right = SCREEN_WIDTH
-        if self.rect.top <= 0:
-            self.rect.top = 0
-        if self.rect.bottom >= SCREEN_HEIGHT:
-            self.rect.bottom = SCREEN_HEIGHT
-
-
-class Wall(pygame.sprite.Sprite):
-    def __init__(self, center_pos: tuple):
-        super().__init__()
-        self.width = DEFAULT_OBJECT_SIZE
-        self.height = DEFAULT_OBJECT_SIZE
-        self.surf = pygame.image.load("images/wall.png").convert_alpha()
-        self.surf.set_colorkey((255, 255, 255), RLEACCEL)
-        self.rect = self.surf.get_rect(center=center_pos)
-
-    @staticmethod
-    def create_centers_of_walls(field_size: tuple, wall_size: tuple):
-        center_width = wall_size[0] + wall_size[0] // 2
-        center_height = wall_size[1] + wall_size[1] // 2
-        centers = []
-
-        while center_height < field_size[1] - wall_size[1]:
-            while center_width < field_size[0] - wall_size[0]:
-                centers.append((center_width, center_height))
-                center_width += 2 * wall_size[0]
-            center_height += 2 * wall_size[1]
-            center_width = wall_size[0] + wall_size[0] // 2
-
-        return centers
-
 
 player = Player()
 
+bombs = pygame.sprite.Group()
 walls = pygame.sprite.Group()
 all_sprites = pygame.sprite.Group()
 all_sprites.add(player)
 
-for wall_center in Wall.create_centers_of_walls(
-    (SCREEN_WIDTH, SCREEN_HEIGHT), (DEFAULT_OBJECT_SIZE, DEFAULT_OBJECT_SIZE)
-):
-    wall = Wall(wall_center)
-    walls.add(wall)
-    all_sprites.add(wall)
+Wall.create_walls(walls=walls, all_sprites=all_sprites)
 
 running = True
 
@@ -112,7 +37,10 @@ while running:
 
     pressed_keys = pygame.key.get_pressed()
 
-    player.update(pressed_keys)
+    player.update(pressed_keys=pressed_keys,
+                  walls=walls,
+                  bombs_group=bombs,
+                  all_sprites_group=all_sprites)
 
     walls.update()
 
@@ -120,11 +48,6 @@ while running:
 
     for sprite in all_sprites:
         screen.blit(sprite.surf, sprite.rect)
-
-    if pygame.sprite.spritecollideany(player, walls):
-        pass
-
-
 
     pygame.display.flip()
     clock.tick(60)
