+++ app/wall.py
@@ -1,0 +1,36 @@
+import pygame
+
+from app.const import DEFAULT_OBJECT_SIZE, SCREEN_WIDTH, SCREEN_HEIGHT
+
+
+class Wall(pygame.sprite.Sprite):
+    def __init__(self, center_pos: tuple):
+        super().__init__()
+        self.width = DEFAULT_OBJECT_SIZE
+        self.height = DEFAULT_OBJECT_SIZE
+        self.surf = pygame.image.load("images/wall.png").convert_alpha()
+        self.rect = self.surf.get_rect(center=center_pos)
+
+    @staticmethod
+    def create_centers_of_walls(field_size: tuple, wall_size: tuple):
+        center_width = wall_size[0] + wall_size[0] // 2
+        center_height = wall_size[1] + wall_size[1] // 2
+        centers = []
+
+        while center_height < field_size[1] - wall_size[1]:
+            while center_width < field_size[0] - wall_size[0]:
+                centers.append((center_width, center_height))
+                center_width += 2 * wall_size[0]
+            center_height += 2 * wall_size[1]
+            center_width = wall_size[0] + wall_size[0] // 2
+
+        return centers
+
+    @staticmethod
+    def create_walls(walls, all_sprites):
+        for wall_center in Wall.create_centers_of_walls(
+                (SCREEN_WIDTH, SCREEN_HEIGHT), (DEFAULT_OBJECT_SIZE, DEFAULT_OBJECT_SIZE)
+        ):
+            wall = Wall(wall_center)
+            walls.add(wall)
+            all_sprites.add(wall)
